+++ NEWS.md
@@ -1,5 +1,7 @@
 # eam (development version)
 
+- Simulation allow more than 1024 data chunks/arrow partitions. Now, it depends on the hard limit of the arrow library and the file system.
+
 ## repo
 
 - Add documentation with pkgdown.